zstd: Add amd64 match length assembly
Copied from the S2 implementation. 5-10% faster.

Replaces #823
